+++ lib/Expression/Ast/UnrepresentableValueNode.php
@@ -1,0 +1,27 @@
+<?php
+
+namespace PhpBench\Expression\Ast;
+
+final class UnrepresentableValueNode implements PhpValue
+{
+    /**
+     * @var mixed
+     */
+    private $value;
+
+    /**
+     * @param mixed $value
+     */
+    public function __construct($value)
+    {
+        $this->value = $value;
+    }
+
+    /**
+     * @return mixed
+     */
+    public function value()
+    {
+        return $this->value;
+    }
+}